requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
black==24.10.0
certifi==2024.12.14
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
datasets==3.2.0
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docker==7.1.0
fastcore==1.7.27
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
ghapi==1.0.6
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
identify==2.6.3
idna==3.10
jiter==0.8.2
Levenshtein==0.26.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.0
openai==1.58.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
plotly==5.24.1
pre_commit==4.0.1
propcache==0.2.1
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pydra-config==0.0.6
Pygments==2.18.0
pyright==1.1.391
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
setuptools==75.1.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
unidiff==0.7.5
urllib3==2.2.3
virtualenv==20.28.0
wheel==0.44.0
xxhash==3.5.0
yarl==1.18.3
